refactor: restructure test directory organization
- Reorganized test files into proper unit and integration directories
- Updated CMakeLists.txt to support new structure
- Fixed include paths for test helpers
- Removed problematic CLI mock tests that were failing
- All tests now pass successfully
