+++ tests/integration/tile_export/test_tile_export.cpp
@@ -25,7 +25,7 @@
 #include <art2img/export.hpp>
 #include <art2img/io.hpp>
 
-#include "../test_helpers.hpp"
+#include "../../test_helpers.hpp"
 
 namespace {
