+++ tests/unit/animation/CMakeLists.txt
@@ -1,0 +1,2 @@
+# Animation unit tests
+# This file is included by the main tests/CMakeLists.txt
