+++ tests/unit/core/error/test_error.cpp
