+++ tests/unit/core/palette/test_palette.cpp
